front: ui-charts: move mousecontext in common folder
  As part of the ui-charts refactoring, we want to move MouseContext to the common folder because it is used in several charts, not just in `SpaceTimeChart`.
