+++ front/ui/ui-charts/src/common/index.ts
@@ -13,3 +13,5 @@ export type {
   PickingElement,
   Point,
 } from './types';
+
+export { MouseContext } from './context';
