+++ front/ui/ui-charts/src/spaceTimeChart/lib/context.ts
@@ -1,10 +1,9 @@
 import { createContext } from 'react';
 
 import type { CanvasContextType } from '../../common/types';
-import type { MouseContextType, SpaceTimeChartContextType } from './types';
+import type { SpaceTimeChartContextType } from './types';
 
-// There are three different contexts because they have very different lifecycles:
-// - MouseContext
+// There are two different contexts because they have very different lifecycles:
 // - SpaceTimeChartContext
 // - SpaceTimeChartCanvasContext
 
@@ -22,13 +21,4 @@ export const SpaceTimeChartCanvasContext = createContext<
 >(
   // That value should never be used, since the context should always be accessed within a provider
   undefined as unknown as CanvasContextType<SpaceTimeChartContextType>
-);
-
-/**
- * This context exports everything necessary to draw the chart.
- * It is updated anytime the scales are updated or the mouse moves.
- */
-export const MouseContext = createContext<MouseContextType>(
-  // That value should never be used, since the context should always be accessed within a provider
-  undefined as unknown as MouseContextType
 );
